brain/pyproject.toml
[project]
name = "biu-brain"
version = "0.1.0"
description = "Brain service for Biu - WhatsApp CV assistant"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "google-genai>=1.0",
    "pydantic>=2.9",
    "jinja2>=3.1",
    "weasyprint>=63.0",
    "structlog>=24.4",
    "python-multipart>=0.0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
